get_messages_to_advance
